add myshop-common and fix divide-service

+++ divide-service/divide-api/src/main/java/com/xcxcxcxcx/myshop/dto/TopicEntity.java
@@ -12,7 +12,7 @@ public class TopicEntity implements Serializable{
 
     private Long topicId;
 
-    private int unitAmount;
+    private double unitAmount;
 
     private Long publisherId;
 
@@ -32,11 +32,11 @@ public class TopicEntity implements Serializable{
         this.topicId = topicId;
     }
 
-    public int getUnitAmount() {
+    public double getUnitAmount() {
         return unitAmount;
     }
 
-    public void setUnitAmount(int unitAmount) {
+    public void setUnitAmount(double unitAmount) {
         this.unitAmount = unitAmount;
     }
